Revert "day7 - using rayon"
This reverts commit bbf362e85d4d832829772be9b0f04a5e33b51065.

+++ src/days/mod.rs
@@ -9,22 +9,22 @@ mod aoc05;
 mod aoc06;
 mod aoc07;
 
-pub fn run(day: i32, first: bool, test: bool) -> Result<i64, io::Error> {
+pub fn run(day: i32, first: bool, test: bool) -> Result<i32, io::Error> {
     let path = resolve_path(day, test);
 
     let output = match (day, first) {
-        (1, true) => aoc01::aoc01_a(path) as i64,
-        (1, false) => aoc01::aoc01_b(path) as i64,
-        (2, true) => aoc02::aoc02_a(path) as i64,
-        (2, false) => aoc02::aoc02_b(path) as i64,
-        (3, true) => aoc03::part_a(path) as i64,
-        (3, false) => aoc03::part_b(path) as i64,
-        (4, true) => aoc04::part_a(path) as i64,
-        (4, false) => aoc04::part_b(path) as i64,
-        (5, true) => aoc05::part_a(path) as i64,
-        (5, false) => aoc05::part_b(path) as i64,
-        (6, true) => aoc06::part_a(path) as i64,
-        (6, false) => aoc06::part_b(path) as i64,
+        (1, true) => aoc01::aoc01_a(path),
+        (1, false) => aoc01::aoc01_b(path),
+        (2, true) => aoc02::aoc02_a(path),
+        (2, false) => aoc02::aoc02_b(path),
+        (3, true) => aoc03::part_a(path),
+        (3, false) => aoc03::part_b(path),
+        (4, true) => aoc04::part_a(path),
+        (4, false) => aoc04::part_b(path),
+        (5, true) => aoc05::part_a(path),
+        (5, false) => aoc05::part_b(path),
+        (6, true) => aoc06::part_a(path),
+        (6, false) => aoc06::part_b(path),
         (7, true) => aoc07::part_a(path),
         (7, false) => aoc07::part_b(path),
         _ => -1,
